+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning v2.0.0](https://semver.org/spec
 - Update encoding script
 - Remove azure pipelines CI/CD
 - Update initialization script
+- Build script fixes encoding
 
 ## [1.3.1] - 2025-01-15